requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
chromadb>=0.4.22
openai>=1.12.0
solana>=0.32.0
solders>=0.20.0
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.26.0
uvicorn>=0.27.0
fastapi>=0.109.0
